refactor all default controls
