--- server/meep-draft/components/index.react.js
@@ -1,1 +1,0 @@
-export FontFamilyControls from './font-family/component.react';
