+++ server/meep-draft/defaultControls/index.react.js
@@ -1,0 +1,8 @@
+export FontFamilyControls from './font-family/component.react';
+export FontSizeControls from './font-size/component.react';
+export TextControls from './text/component.react';
+export LinkControls from './link/component.react';
+export BlockControls from './block/component.react';
+export ColorControls from './color/component.react';
+export BackgroundControls from './background/component.react';
+export ContentControls from './content/component.react';
