+++ server/meep-draft/draft-text.react.js
@@ -1,7 +1,16 @@
 import React, { PropTypes, Component } from 'react';
 import ReactDOM from 'react-dom';
 // components load
-import * as DefaultControlComponents from './components/index.react';
+import {
+  FontFamilyControls,
+  FontSizeControls,
+  TextControls,
+  LinkControls,
+  BlockControls,
+  ColorControls,
+  BackgroundControls,
+  ContentControls
+} from './defaultControls/index.react';
 //
 import merge from './lib/merge.js'
 import styles from './draft-text.style'
@@ -17,19 +26,6 @@ import Draft, {
   ContentState,
   SelectionState,
   CharacterMetadata } from 'draft-js';
-
-import camelcase from 'camelcase';
-
-//type-core
-import BLOCK_TYPES from './draft-type-core/block'
-import {
-  FONTSIZESTYLE,
-  TEXTSTYLE,
-  COLORSTYLE,
-  BACKGROUNDCOLORSTYLE,
-  ALIGNSTYLE,
-  FONTFAMILYSTYLE
-} from './draft-type-core/inline'
 //custom-core
 import {
   FONTSIZE,
@@ -465,21 +461,7 @@ export default class DraftText extends Component {
     const rootStyle = this.checkRootStyle ? (this.props.editorStyle.root) : ({})
     const rootControlStyle = this.checkRootControlStyle ? (this.props.editorStyle['root-control']) : ({})
     const rootInputStyle = this.checkRootInputStyle ? (this.props.editorStyle['root-input']) : ({})
-    // let render = [];
 
-    // this.ControlsComponentsRender = () => {
-    //   for(let c in DefaultControlComponents) {
-    //     let ControlsComponent = DefaultControlComponents[c];
-    //     render.push(
-    //       <ControlsComponent
-    //         key={c}
-    //         editorState={editorState}
-    //         customStyle={rootControlStyle}
-    //       />
-    //     )
-    //   }
-    // }
-    // this.ControlsComponentsRender();
     const {
       controls
     } = this.defaultSetting
@@ -544,6 +526,7 @@ export default class DraftText extends Component {
         groupControls={controls.content}
       />
     ) : null
+
     const controlsComponentEditor = this.props.readOnly ? null : (
       <div>
         {fontFamilyControls}
@@ -598,568 +581,3 @@ const Link = (props) => {
     </a>
   );
 };
-
-class FontFamilyControls extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      onOpen: false
-    }
-
-    this._onOpen = () => {
-      this.setState({
-        onOpen: !this.state.onOpen
-      })
-    }
-  }
-  render() {
-    let currentStyle = this.props.editorState.getCurrentInlineStyle();
-    let fontFamily = 'Arial';
-    let itemMap = (
-      ['Arial', 'Arial Black', 'Comic Sans MS', 'Courier', 'Courier New', '標楷體', 'Helvetica','Impace','Lucida Grande','Lucida Sans','微軟正黑體', 'Monospace','新細明體','Sans Serif','Serif','Tahoma','Times','Times New Roman','Verdana'].map((family, idx) => {
-          if(currentStyle.has(FONTFAMILYSTYLE[idx].style)) {
-            fontFamily = (FONTFAMILYSTYLE[idx].style)
-          }
-          return (
-            <SelectFamilyItem
-              active={currentStyle.has(FONTFAMILYSTYLE[idx].style)}
-              key={`font_family_button_${idx}`}
-              size={family}
-              style={FONTFAMILYSTYLE[idx].style}
-              onToggle={this.props.onToggle}
-              onOpen={this._onOpen}
-            />
-          )
-      })
-    )
-    let items = this.state.onOpen ? (
-      <div
-        style={styles.meepEditorSelectItemBox}
-      >
-        {itemMap}
-      </div>
-    ) : ( null )
-    let customControlStyle = this.props.customStyle
-    return (
-      <div
-        style={merge(
-          styles.meepEditorInline,
-          styles.meepEditorInlineFontFamily
-        )}
-      >
-        <div
-          style={merge(styles.meepEditorSelectMainBox,
-                       customControlStyle,
-                       this.state.onOpen?styles.meepEditorSelectMainBoxOpen:null)}
-        >
-          <div
-            onClick={this._onOpen}
-          >
-            <span
-              style={styles.meepEditorSelectBoxLabel}
-            >{fontFamily}</span>
-            <i
-              style={styles.meepEditorSelectBoxIcon}
-              className="fa fa-caret-down">
-            </i>
-          </div>
-          {items}
-        </div>
-      </div>
-    );
-  }
-}
-
-class FontSizeControls extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      onOpen: false,
-    }
-
-    this._onOpen = () => {
-      this.setState({
-        onOpen: !this.state.onOpen
-      })
-    }
-  }
-  render() {
-    let currentStyle = this.props.editorState.getCurrentInlineStyle();
-    let fontSize = 10;
-    let itemMap = (
-      [10, 13, 16, 20, 24, 28, 32].map((size, idx) => {
-        if(currentStyle.has(FONTSIZESTYLE[idx].style)) {
-          fontSize = (FONTSIZESTYLE[idx].style).split('-')[1]
-        }
-        return (
-          <SelectItem
-            active={currentStyle.has(FONTSIZESTYLE[idx].style)}
-            key={`font_size_button_${idx}`}
-            size={size}
-            style={FONTSIZESTYLE[idx].style}
-            onToggle={this.props.onToggle}
-            onOpen={this._onOpen}
-          />
-        )
-      })
-    )
-    let items = this.state.onOpen ? (
-      <div
-        style={styles.meepEditorSelectItemBox}
-      >
-        {itemMap}
-      </div>
-    ) : ( null )
-    let customControlStyle = this.props.customStyle
-    return (
-      <div
-        style={styles.meepEditorInline}
-      >
-        <div
-          style={merge(styles.meepEditorSelectMainBox,
-                       customControlStyle,
-                       this.state.onOpen?styles.meepEditorSelectMainBoxOpen:{})}
-        >
-          <div
-            onClick={this._onOpen}
-          >
-            <span
-              style={styles.meepEditorSelectBoxLabel}
-            >{fontSize} px</span>
-            <i
-              style={styles.meepEditorSelectBoxIcon}
-              className="fa fa-caret-down">
-            </i>
-          </div>
-          {items}
-        </div>
-      </div>
-    );
-  }
-}
-
-class SelectItem extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      hover: false
-    }
-
-    this.onToggle = (e) => {
-      e.preventDefault();
-      this.props.onToggle(this.props.style);
-    }
-
-    this._onHover = () => {
-      this.setState({
-        hover: true
-      })
-    }
-    this._onLeave = () => {
-      this.setState({
-        hover: false
-      })
-    }
-  }
-  render() {
-    return (
-      <span
-        style={merge(styles.meepEditorSelectItem,
-                     this.state.hover?styles.meepEditorSelectItemHover:null,
-                     this.props.active?styles.meepEditorActiveButton:null
-                    )}
-        onMouseOver={this._onHover}
-        onMouseLeave={this._onLeave}
-        onClick={(event)=>{
-          this.onToggle(event)
-          this.props.onOpen()
-        }}
-      >
-        {this.props.size} px
-      </span>
-    );
-  }
-}
-
-class SelectFamilyItem extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      hover: false
-    }
-
-    this.onToggle = (e) => {
-      e.preventDefault();
-      this.props.onToggle(this.props.style);
-    }
-
-    this._onHover = () => {
-      this.setState({
-        hover: true
-      })
-    }
-    this._onLeave = () => {
-      this.setState({
-        hover: false
-      })
-    }
-  }
-  render() {
-    return (
-      <span
-        style={merge(styles.meepEditorSelectFamilyItem,
-                     this.props.active?styles.meepEditorActiveButton:null,
-                     this.state.hover?styles.meepEditorSelectItemHover:null)
-              }
-        onMouseOver={this._onHover}
-        onMouseLeave={this._onLeave}
-        onClick={(event)=>{
-          this.onToggle(event)
-          this.props.onOpen()
-        }}
-      >
-        {this.props.size}
-      </span>
-    );
-  }
-}
-
-const TextControls = (props) => {
-  const currentStyle = props.editorState.getCurrentInlineStyle();
-  const {
-    groupControls
-  } = props
-  const button = (TEXTSTYLE.map((type, index) => {
-    if (!groupControls[type.style]) return
-    return (
-      <StyleButton
-        key={`text_button_${index}`}
-        active={currentStyle.has(type.style)}
-        label={<i className={type.label}></i>}
-        style={type.style}
-        onToggle={props.onToggle}
-      /> )
-  }))
-  return (
-    <div
-      style={
-        merge(styles.controls,
-              styles.meepEditorInline,
-             )
-      }
-    >
-      {button}
-    </div>
-  )
-}
-
-const LinkControls = (props) => {
-  const {
-    onHandlLink,
-    groupControls
-  } = props
-  const links = [];
-
-  groupControls.set ? links.push(
-    <span
-      key={`link-2`}
-      style={merge(styles.meepEditorDefaultColor, styles.meepEditorDefaultButton)}
-      onClick={(e) => {
-        onHandlLink(e, 'addLink')
-      }}
-    >
-      <i className="fa fa-link"></i>
-    </span>
-  ) : null
-
-  groupControls.unset ? links.push(
-    <span
-      key={`link-1`}
-      style={merge(styles.meepEditorDefaultColor, styles.meepEditorDefaultButton)}
-      onClick={(e) => {
-        onHandlLink(e, 'removeLink')
-      }}
-    >
-      <i className="fa fa-unlink"></i>
-    </span>
-  ) : null
-
-  return (
-    <span>
-      {links}
-    </span>
-  )
-}
-
-const BlockControls = (props) => {
-  let {
-    editorState,
-    groupControls
-  } = props
-  const selection = editorState.getSelection();
-  const blockType = editorState.getCurrentContent()
-                               .getBlockForKey(selection.getStartKey())
-                               .getType();
-  let button = (BLOCK_TYPES.map((type, index) => {
-    if (!groupControls[camelcase(type.style)]) return
-    return (
-      <StyleButton
-        key={`block_button_${index}`}
-        active={type.style === blockType}
-        label={<i className={type.label}></i>}
-        style={type.style}
-        onToggle={props.onToggle}
-      /> )
-  }))
-  return (
-    <div
-      style={styles.meepEditorInline}
-    >
-      {button}
-    </div>
-  )
-}
-
-class ColorControls extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      isOpen: false
-    }
-
-    this._onOpen = () => {
-      this.setState({
-        isOpen: !this.state.isOpen
-      })
-    }
-  }
-  render() {
-    let {
-      editorState
-    } = this.props
-    let currentStyle = editorState.getCurrentInlineStyle();
-    let button = this.state.isOpen ? (COLORSTYLE.map((type, idx) => {
-      return (
-        <ColorButton
-          key={`color_button_${idx}`}
-          active={currentStyle.has(type.style)}
-          label={type.label}
-          style={type.style}
-          onToggle={this.props.onToggle}
-        />)
-    })) : ( null )
-    return (
-      <div
-        style={styles.meepEditorInline}
-      >
-        <div
-          style={merge(styles.meepEditorDefaultButton, styles.meepEditorActionSelect)}
-          onClick={this._onOpen}
-        >
-          <i className="fa fa-font"></i>
-        </div>
-        <div
-          style={styles.meepEditorActiveColorBox}
-        >
-          {button}
-        </div>
-      </div>
-    )
-  }
-}
-
-class BackgroundControls extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      isOpen: false
-    }
-
-    this._onOpen = () => {
-      this.setState({
-        isOpen: !this.state.isOpen
-      })
-    }
-  }
-  render() {
-    let {
-      editorState
-    } = this.props
-    let currentStyle = editorState.getCurrentInlineStyle();
-    let button = this.state.isOpen ? (BACKGROUNDCOLORSTYLE.map((type, idx) => {
-      return (
-        <BackgroundButton
-          key={`background_button_${idx}`}
-          active={currentStyle.has(type.style)}
-          label={type.label}
-          style={type.style}
-          onToggle={this.props.onToggle}
-        />)
-    })) : ( null )
-    return (
-      <div
-        style={styles.meepEditorInline}
-      >
-        <div
-          style={merge(styles.meepEditorDefaultButton, styles.meepEditorActionSelect)}
-          onClick={this._onOpen}
-        >
-          <i className="fa fa-cog"></i>
-        </div>
-        <div
-          style={styles.meepEditorActiveBackgroundBox}
-        >
-          {button}
-        </div>
-      </div>
-    );
-  }
-}
-
-const ContentControls = (props) => {
-  const {
-    groupControls
-  } = props
-
-  const content = [];
-  
-  groupControls.redo ? (
-    content.push(
-      <ContentButton
-        key={`content-1`}
-        label={<i className="fa fa-repeat"></i>}
-        doAction="redo"
-        editorState={props.editorState}
-        onDoHandle={props.onDoHandle}
-      />
-    )
-  ) : null
-
-  groupControls.undo ? (
-    content.push(
-      <ContentButton
-        key={`content-2`}
-        label={<i className="fa fa-undo"></i>}
-        doAction="undo"
-        editorState={props.editorState}
-        onDoHandle={props.onDoHandle}
-      />
-    )
-  ) : null
-
-  return (
-    <div
-      style={styles.meepEditorInline}
-    >
-      {content}
-    </div>
-  );
-}
-
-class ContentButton extends Component {
-  constructor(props) {
-    super(props);
-  }
-  render() {
-    return (
-      <span
-        style={merge(styles.meepEditorDefaultColor, styles.meepEditorDefaultButton)}
-        onMouseDown={() => {
-          this.props.onDoHandle(this.props.editorState, this.props.doAction)
-        }}
-      >
-        {this.props.label}
-      </span>
-    );
-  }
-}
-
-class StyleButton extends Component {
-  constructor(props) {
-    super(props);
-    this.onToggle = (e) => {
-      e.preventDefault();
-      this.props.onToggle(this.props.style);
-    };
-  }
-  render() {
-    let style;
-    let labelColor;
-    if (this.props.active) {
-      style = styles.meepEditorActiveButton;
-      labelColor = '#437A82'
-    } else {
-      style = styles.styleButton;
-      labelColor = '#59bcc9'
-    }
-
-    return (
-      <span
-        style={merge(style, styles.meepEditorDefaultButton)}
-        labelColor={labelColor}
-        onMouseDown={this.onToggle}
-      >
-        {this.props.label}
-      </span>
-    );
-  }
-}
-
-class ColorButton extends Component {
-  constructor(props) {
-    super(props);
-    this.onToggle = (e) => {
-      e.preventDefault();
-      this.props.onToggle(this.props.style);
-    };
-  }
-  render() {
-    let style;
-    let labelColor;
-    if (this.props.active) {
-      style = styles.meepEditorActiveColorButton;
-    } else {
-      style = styles.styleButton;
-    }
-
-    return (
-      <span
-        style={merge(styles.meepEditorDefaultColorButton, style, {backgroundColor: this.props.style})}
-        onMouseDown={this.onToggle}
-      >
-      </span>
-    );
-  }
-}
-
-class BackgroundButton extends Component {
-  constructor(props) {
-    super(props);
-    this.onToggle = (e) => {
-      e.preventDefault();
-      this.props.onToggle(this.props.style);
-    };
-  }
-  render() {
-    let style;
-    let labelColor;
-    if (this.props.active) {
-      style = styles.meepEditorActiveColorButton;
-    } else {
-      style = styles.styleButton;
-    }
-    return (
-      <span
-        style={merge(styles.meepEditorDefaultColorButton, style, {backgroundColor: this.props.style.split('-')[1]})}
-        onMouseDown={this.onToggle}
-      >
-      </span>
-    );
-  }
-}
